10_projects/paper_soflow/requirements.txt
# SoFlow: Solution Flow Models for One-Step Generative Modeling
# Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Diffusion/Flow utilities
einops>=0.6.0
timm>=0.9.0

# Training utilities
accelerate>=0.20.0
wandb>=0.15.0
tqdm>=4.65.0

# Image processing
pillow>=9.5.0
opencv-python>=4.7.0

# Configuration
omegaconf>=2.3.0
hydra-core>=1.3.0

# Visualization
matplotlib>=3.7.0

# Optional: For ImageNet dataset
datasets>=2.12.0

